# perf: add faces to Surface one by one in Surface._setup_in_uv_space