@chopan050 2022-12-26
UPDATE: I made a new PR which, if accepted, would render this PR useless. Please review it here: https://github.com/ManimCommunity/manim/pull/3092